+++ CHANGELOG.md
@@ -81,3 +81,4 @@ has been improved.
   He is now also a maintainer.
 - Added image used by Stylix into the host variables file.
 - Made printing and NFS variables so they can be easily toggled between hosts.
+- Added waybar styling choice.